Addition of findByStatusPetAPICall flow - Added findByStatus API call in UpdatePetsDetailsTest - Added seperate API with findByStatusFlow

+++ src/main/java/common/pets/Pet.java
@@ -1,11 +1,11 @@
 package common.pets;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Getter
@@ -15,7 +15,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Pet {
 
-    private int id;
+    private double id;
     private String name;
     private String status;
 
@@ -23,7 +23,7 @@ public class Pet {
     private List<String> photoUrls = null;
     private List<Tag> tags = null;
 
-    public Pet(int id, String name, String status, Category category, List<String> photoUrls, List<Tag> tags){
+    public Pet(double id, String name, String status, Category category, List<String> photoUrls, List<Tag> tags){
         this.id = id;
         this.name = name;
         this.status = status;
